requirements.txt
# Core AI frameworks
google-generativeai>=0.3.0
crewai>=0.70.0
langgraph>=0.0.40
langchain>=0.1.0
langchain-google-genai>=0.0.6
python-a2a>=0.5.0

# Web framework
streamlit>=1.30.0
flask>=3.0.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.0

# Data handling
pandas>=2.0.0
datetime

# Additional
httpx>=0.26.0

langchain-groq>=0.0.5
